pyproject.toml
[project]
name = "test-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai",
    "livekit-plugins-assemblyai",
    "livekit-plugins-cartesia",
    "python-dotenv>=1.2.1",
]
